# Add quantity to inventory units and split on demand