--- a/.gitignore
+++ b/.gitignore
@@ -1,6 +1,7 @@
 \#*
 **/*.db
 **/*.log
+**/*.byebug_history
 *~
 .#*
 .DS_Store