--- a/core/app/models/spree/calculator/returns/default_refund_amount.rb
+++ b/core/app/models/spree/calculator/returns/default_refund_amount.rb
@@ -21,8 +21,8 @@ module Spree
       end
 
       def percentage_of_order_total(return_item, inventory_unit)
-       return 0.0 if inventory_unit.order.pre_tax_item_amount.zero?
-       weighted_line_item_pre_tax_amount(return_item, inventory_unit) / inventory_unit.order.pre_tax_item_amount
+        return 0.0 if inventory_unit.order.pre_tax_item_amount.zero?
+        weighted_line_item_pre_tax_amount(return_item, inventory_unit) / inventory_unit.order.pre_tax_item_amount
       end
 
       def weighted_line_item_pre_tax_amount(return_item, inventory_unit)
@@ -32,6 +32,6 @@ module Spree
       def percentage_of_line_item(return_item, inventory_unit)
         return_item.return_quantity / BigDecimal.new(inventory_unit.line_item.quantity)
       end
-   end
- end
+    end
+  end
 end