--- a/core/app/models/spree/order_inventory.rb
+++ b/core/app/models/spree/order_inventory.rb
@@ -25,9 +25,7 @@ module Spree
 
           shipment = determine_target_shipment unless shipment
           add_to_shipment(shipment, quantity)
-        elsif units_count == line_item.quantity && !line_item.changed?
-          remove(units_count, shipment)
-        elsif units_count > line_item.quantity
+        elsif (units_count > line_item.quantity) || (units_count == line_item.quantity && !line_item.changed?)
           remove(units_count, shipment)
         end
       end