--- a/core/app/models/spree/order_inventory.rb
+++ b/core/app/models/spree/order_inventory.rb
@@ -33,86 +33,86 @@ module Spree
 
     private
 
-      def remove(units_count, shipment = nil)
-        quantity = set_quantity_to_remove(units_count)
+    def remove(units_count, target_shipment = nil)
+      quantity = set_quantity_to_remove(units_count)
 
-        if shipment.present?
-          remove_from_shipment(shipment, quantity)
-        else
-          order.shipments.each do |shipment|
-            break if quantity == 0
-            quantity -= remove_from_shipment(shipment, quantity)
+      if target_shipment.present?
+        remove_from_shipment(target_shipment, quantity)
+      else
+        order.shipments.each do |shipment|
+          break if quantity.zero?
+          quantity -= remove_from_shipment(shipment, quantity)
+        end
+      end
+    end
+
+    def set_quantity_to_remove(units_count)
+      if (units_count - line_item.quantity).zero?
+        line_item.quantity
+      else
+        units_count - line_item.quantity
+      end
+    end
+
+    # Returns either one of the shipment:
+    #
+    # first unshipped that already includes this variant
+    # first unshipped that's leaving from a stock_location that stocks this variant
+    def determine_target_shipment
+      target_shipment = order.shipments.detect do |shipment|
+        shipment.ready_or_pending? && shipment.include?(variant)
+      end
+
+      target_shipment || order.shipments.detect do |shipment|
+        shipment.ready_or_pending? && variant.stock_location_ids.include?(shipment.stock_location_id)
+      end
+    end
+
+    def add_to_shipment(shipment, quantity)
+      if variant.should_track_inventory?
+        on_hand, back_order = shipment.stock_location.fill_status(variant, quantity)
+
+        on_hand.times { shipment.set_up_inventory('on_hand', variant, order, line_item) }
+        back_order.times { shipment.set_up_inventory('backordered', variant, order, line_item) }
+      else
+        quantity.times { shipment.set_up_inventory('on_hand', variant, order, line_item) }
+      end
+
+      # adding to this shipment, and removing from stock_location
+      if order.completed?
+        shipment.stock_location.unstock(variant, quantity, shipment)
+      end
+
+      quantity
+    end
+
+    def remove_from_shipment(shipment, quantity)
+      return 0 if quantity.zero? || shipment.shipped?
+
+      shipment_units = shipment.inventory_units_for_item(line_item, variant).reject(&:shipped?).sort_by(&:state)
+
+      removed_quantity = 0
+
+      shipment_units.each do |inventory_unit|
+        inventory_unit.quantity.times do
+          break if removed_quantity == quantity
+          if inventory_unit.quantity > 1
+            inventory_unit.decrement(:quantity)
+          else
+            inventory_unit.destroy
           end
+          removed_quantity += 1
         end
       end
 
-      def set_quantity_to_remove(units_count)
-        if (units_count - line_item.quantity).zero?
-          line_item.quantity
-        else
-          units_count - line_item.quantity
-        end
+      shipment.destroy if shipment.inventory_units.sum(:quantity).zero?
+
+      # removing this from shipment, and adding to stock_location
+      if order.completed?
+        shipment.stock_location.restock variant, removed_quantity, shipment
       end
 
-      # Returns either one of the shipment:
-      #
-      # first unshipped that already includes this variant
-      # first unshipped that's leaving from a stock_location that stocks this variant
-      def determine_target_shipment
-        shipment = order.shipments.detect do |shipment|
-          shipment.ready_or_pending? && shipment.include?(variant)
-        end
-
-        shipment ||= order.shipments.detect do |shipment|
-          shipment.ready_or_pending? && variant.stock_location_ids.include?(shipment.stock_location_id)
-        end
-      end
-
-      def add_to_shipment(shipment, quantity)
-        if variant.should_track_inventory?
-          on_hand, back_order = shipment.stock_location.fill_status(variant, quantity)
-
-          on_hand.times { shipment.set_up_inventory('on_hand', variant, order, line_item) }
-          back_order.times { shipment.set_up_inventory('backordered', variant, order, line_item) }
-        else
-          quantity.times { shipment.set_up_inventory('on_hand', variant, order, line_item) }
-        end
-
-        # adding to this shipment, and removing from stock_location
-        if order.completed?
-          shipment.stock_location.unstock(variant, quantity, shipment)
-        end
-
-        quantity
-      end
-
-      def remove_from_shipment(shipment, quantity)
-        return 0 if quantity == 0 || shipment.shipped?
-
-        shipment_units = shipment.inventory_units_for_item(line_item, variant).reject(&:shipped?).sort_by(&:state)
-
-        removed_quantity = 0
-
-        shipment_units.each do |inventory_unit|
-          inventory_unit.quantity.times do
-            break if removed_quantity == quantity
-            if inventory_unit.quantity > 1
-              inventory_unit.decrement(:quantity)
-            else
-              inventory_unit.destroy
-            end
-            removed_quantity += 1
-          end
-        end
-
-        shipment.destroy if shipment.inventory_units.sum(:quantity).zero?
-
-        # removing this from shipment, and adding to stock_location
-        if order.completed?
-          shipment.stock_location.restock variant, removed_quantity, shipment
-        end
-
-        removed_quantity
-      end
+      removed_quantity
+    end
   end
 end