--- a/core/app/models/spree/shipment.rb
+++ b/core/app/models/spree/shipment.rb
@@ -172,11 +172,11 @@ module Spree
         units.group_by(&:line_item_id).map do |line_item_id, units|
 
           states = {}
-          units.group_by(&:state).each { |state, iu| states[state] = iu.size }
+          units.group_by(&:state).each { |state, iu| states[state] = iu.sum(&:quantity) }
 
           line_item = units.first.line_item
           variant = units.first.variant
-          ManifestItem.new(line_item, variant, units.length, states)
+          ManifestItem.new(line_item, variant, units.sum(&:quantity), states)
         end
       end.flatten
     end