… api/spec/controllers/spree/api/v1/line_items_controller_spec.rb
… core/app/models/spree/stock/availability_validator.rb
… core/app/models/spree/stock/content_item.rb
… core/app/models/spree/stock/coordinator.rb
… core/app/models/spree/stock/inventory_unit_builder.rb
… core/app/models/spree/stock/package.rb
… core/app/models/spree/stock/packer.rb
… core/app/models/spree/stock/prioritizer.rb
… core/app/models/spree/stock/splitter/weight.rb
… core/app/models/spree/stock_item.rb
… core/app/models/spree/stock_location.rb
… core/config/locales/en.yml
… core/db/migrate/20161125065505_add_quantity_to_inventory_units.rb
… core/db/migrate/20170119122701_add_original_return_item_id_to_spree_inventory_units.rb
… core/lib/spree/core/importer/order.rb
… core/lib/tasks/exchanges.rake
… core/spec/lib/spree/core/importer/order_spec.rb
… core/spec/lib/tasks/exchanges_spec.rb
… core/spec/models/spree/calculator/refunds/default_refund_amount_spec.rb
… core/spec/models/spree/line_item_spec.rb
… core/spec/models/spree/order_inventory_spec.rb
… core/spec/models/spree/return_authorization_spec.rb
… core/spec/models/spree/return_item_spec.rb
… core/spec/models/spree/shipment_spec.rb
… core/spec/models/spree/stock/availability_validator_spec.rb
… core/spec/models/spree/stock/coordinator_spec.rb
… core/spec/models/spree/stock/inventory_unit_builder_spec.rb
… core/spec/models/spree/stock/prioritizer_spec.rb
… core/spec/models/spree/stock_item_spec.rb
… frontend/spec/features/checkout_spec.rb